% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_ba.R
\name{get_data_ba}
\alias{get_data_ba}
\title{get_data_ba}
\usage{
get_data_ba(url)
}
\arguments{
\item{url}{string}
}
\value{
Tibble
}
\description{
get_data_ba
}
\examples{
\dontrun{
get_data_ba('https://www.encurtador.com.br/rtvOT')
}
}
