% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_mg.R
\name{get_data_mg}
\alias{get_data_mg}
\title{Get Data Mg}
\usage{
get_data_mg(url)
}
\arguments{
\item{url}{string}
}
\value{
tibble
}
\description{
Get Data Mg
}
\examples{

\dontrun{
get_data_mg('http://www.dig.org.br/noticia/Julho-2021-distrito-do-gorutuba/272')
}
}
