% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_pb.R
\name{get_data_pb}
\alias{get_data_pb}
\title{Get Data Pb}
\usage{
get_data_pb(epoch = as.POSIXct(Sys.time()) \%>\% as.integer())
}
\arguments{
\item{epoch}{integer}
}
\value{
tibble
}
\description{
Get Data Pb
}
\examples{
\dontrun{get_data_pb()}

}
