% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_pe.R
\name{get_data_pe}
\alias{get_data_pe}
\title{Get Data Pe}
\usage{
get_data_pe(url_pe)
}
\arguments{
\item{url_pe}{string}
}
\value{
tibble
}
\description{
Get Data Pe
}
\examples{
\dontrun{get_data_pe('https://www.encurtador.com.br/giwE6')
}

}
