% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_se.R
\name{pdf_to_se}
\alias{pdf_to_se}
\title{Title}
\usage{
pdf_to_se(texto_reservatorio_se, ano_arquivo)
}
\arguments{
\item{texto_reservatorio_se}{string}

\item{ano_arquivo}{string}
}
\value{
tibble
}
\description{
Title
}
